ANPR Web App/requirements.txt
flask
numpy
opencv-python
matplotlib
pytesseract
gunicorn
click
Flask
gunicorn
itsdangerous
Jinja2
MarkupSafe
Werkzeug
Flask-Cors
psycopg
numpy
flask_cors
flask
werkzeug
scikit-image
pillow